remove top-of-file LICENSE header
we usually avoid this boilerplate,
there is no much point in adding it,
all files are covered by the LICENSE file.

this is also common practise in most other rust projects,
so it cannot be that wrong :)

for dedicated contributors,
the git history should be good enough;
emails addresses and names usually do not age well in headers.